platform/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "openclaw-platform"
version = "1.0.0"
description = "SilentLake Multi-Agent Platform - OpenClaw 二开扩展"
authors = [{name = "SilentLake Team"}]
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "apscheduler>=3.10.0",
    "redis>=5.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "anthropic>=0.26.0",
    "psutil>=5.9.0",
]

[project.scripts]
oc-platform = "main:cli"

[tool.setuptools.packages.find]
where = ["."]
